Move text_fg_override code into vertex shader
This should be a ~100x speedup (depending on pixels per cell) for using this code. Note that it also means
that color emoji do not have their colors overridden. This makes sense
since overriding color sin them is not desired anyway and even otherwise
the performance benefits are too large.