kitty +open: Ask for confirmation when running executables
Workaround for badly designed program that try to open links in
untrusted documents that point to executables on the local system.
They should really not be doing that, but, since such programs do exist,
let's be safe by default.

Users that use better designed programs and want to disable the
confirmations can do so by writing their own launch-actions.conf