options: add copy_or_noop, make default
Add a "copy_or_noop" option which copies the text, and if there is no
selected text passes through the key. This allows applications to
implement native text selection but also have cmd+c still function as a
copy key since they can now receive it.

Make this the default on macos. This should have very little impact on
users because when no text is selected, the application can
(potentially) handle the key.