Revert "Workaround for latest Apple regression in Tahoe"
This reverts commit b66dec097da491511ed7ec13696817b66d8d571c.

Commit e159fa3ed9521d80aee9da771a9d0faf3e1715c4 fixes the original
problem of "ghost windows" that b66dec0 was trying to solve. (I have no
idea how or why it does.)

One side-effect of the workaround from b66dec0 is that it causes closed
windows to be moved to the (0,0) before being destroyed. This means that
when creating new windows, they are placed on the left-most display,
instead of the display in which the previously closed window was on as
you would expect. Reverting this kludge fixes this bug.